+++ frontend/components/create-review/review-form.tsx
@@ -9,22 +9,19 @@ import HCaptcha from '@hcaptcha/react-hcaptcha'
 import MaliciousStringAlert from '../alerts/MaliciousStringAlert'
 import RatingsRadio from './ratings-radio'
 import SuccessModal from './success-modal'
-import countries from '@/util/countries.json'
+import countries from '@/util/countries/countries.json'
 import {postcodeValidator} from 'postcode-validator'
-import provinces from '@/util/provinces.json'
-import regions from '@/util/regions.json'
-import states from '@/util/states.json'
-import territories from '@/util/territories.json'
-import nz_provinces from '@/util/nz-provinces.json'
+import provinces from '@/util/countries/canada/provinces.json'
+import regions from '@/util/countries/unitedKingdom/regions.json'
+import states from '@/util/countries/unitedStates/states.json'
+import territories from '@/util/countries/australia/territories.json'
+import nz_provinces from '@/util/countries/newZealand/nz-provinces.json'
 import {useTranslation} from 'react-i18next'
+import {country_codes} from '@/util/helpers/getCountryCodes'
 
 //This components will hold the review form and it's data handling logic
 //Completed reviews should be sent to the backend with a success confirmation for the user (maybe need a Modal?)
 //Once completed, it should give an option to reset the form for another review or direct Client to Reviews page
-
-const country_codes = Object.keys(countries).filter(
-	(c) => c === 'CA' || c === 'US' || c === 'GB' || c === 'AU' || c === 'NZ',
-)
 
 const siteKey = process.env.NEXT_PUBLIC_HCPATCHA_SITE_KEY as string
