+++ frontend/util/helpers/fetcher.ts
@@ -1,0 +1,1 @@
+export const fetcher = (url: string) => fetch(url).then((r) => r.json())
