+++ frontend/util/helpers/helper-functions.ts
@@ -1,4 +1,4 @@
-import {Options} from './interfaces'
+import {Options} from '../interfaces/interfaces'
 
 export function classNames(...classes: Array<string>) {
 	return classes.filter(Boolean).join(' ')
